+++ src/api/dm/addon/math/curves/plane_curves.d
@@ -1,6 +1,6 @@
 module api.dm.addon.math.curves.plane_curves;
 
-import api.math.geom2.vec2 : Vec2d;
+import api.math.geom2.vec2 : Vec2f;
 
 import Math = api.dm.math;
 
@@ -31,18 +31,18 @@ void onPointStep(float step, float minValueInclusive, float maxValueInclusive, s
     }
 }
 
-void witchOfAgnesi(scope bool delegate(Vec2d) onPointIsContinue, float radius = 50, float step = 0.01)
+void witchOfAgnesi(scope bool delegate(Vec2f) onPointIsContinue, float radius = 50, float step = 0.01)
 {
     assert(onPointIsContinue);
 
     onPointStep(step, -(Math.PI / 2) + step, Math.PI / 2 - step, (dt) {
         const x = radius * Math.tan(dt);
         const y = radius * (Math.cos(dt) ^^ 2);
-        return onPointIsContinue(Vec2d(x, y));
+        return onPointIsContinue(Vec2f(x, y));
     });
 }
 
-void bicorn(scope bool delegate(Vec2d) onPointIsContinue, float radius = 50, float thetaRad = 0.01, size_t dots = 500, float step = 1.0)
+void bicorn(scope bool delegate(Vec2f) onPointIsContinue, float radius = 50, float thetaRad = 0.01, size_t dots = 500, float step = 1.0)
 {
     assert(onPointIsContinue);
 
@@ -54,11 +54,11 @@ void bicorn(scope bool delegate(Vec2d) onPointIsContinue, float radius = 50, flo
             /
             (3 + (Math.sin(thetaRad * dt) ^^ 2))
         );
-        return onPointIsContinue(Vec2d(x, y));
+        return onPointIsContinue(Vec2f(x, y));
     });
 }
 
-void cardioid(scope bool delegate(Vec2d) onPointIsContinue, float radius = 10, float step = 0.01)
+void cardioid(scope bool delegate(Vec2f) onPointIsContinue, float radius = 10, float step = 0.01)
 {
     assert(onPointIsContinue);
 
@@ -66,12 +66,12 @@ void cardioid(scope bool delegate(Vec2d) onPointIsContinue, float radius = 10, f
     onPointStep(step, 0, Math.PI * 2, (angle) {
         auto x = (2 * radius) * (1 - Math.cos(angle)) * Math.cos(angle);
         auto y = (2 * radius) * (1 - Math.cos(angle)) * Math.sin(angle);
-        onPointIsContinue(Vec2d(x, y));
+        onPointIsContinue(Vec2f(x, y));
         return true;
     });
 }
 
-void lemniscateBernoulli(scope bool delegate(Vec2d) onPointIsContinue, float distance = 10, float step = 0.01)
+void lemniscateBernoulli(scope bool delegate(Vec2f) onPointIsContinue, float distance = 10, float step = 0.01)
 {
     assert(onPointIsContinue);
 
@@ -83,33 +83,33 @@ void lemniscateBernoulli(scope bool delegate(Vec2d) onPointIsContinue, float dis
         auto y = (distance * Math.sqrt(2.0) * Math.sin(angle) * Math.cos(angle)) / (
             1 + Math.pow(Math.sin(angle), 2));
 
-        return onPointIsContinue(Vec2d(x, y));
+        return onPointIsContinue(Vec2f(x, y));
         return true;
     });
 }
 
-void strophoid(scope bool delegate(Vec2d) onPointIsContinue, float phi, float step = 0.01, float scale = 1.0)
+void strophoid(scope bool delegate(Vec2f) onPointIsContinue, float phi, float step = 0.01, float scale = 1.0)
 {
     assert(onPointIsContinue);
 
     onPointStep(step, 0, Math.PI * 2, (angle) {
         const r = -((phi * Math.cos(2 * angle)) / Math.cos(angle)) * scale;
-        return onPointIsContinue(Vec2d.fromPolarRad(angle, r));
+        return onPointIsContinue(Vec2f.fromPolarRad(angle, r));
     });
 }
 
-void foliumOfDescartes(scope bool delegate(Vec2d) onPointIsContinue, float phi, float step = 0.01, float scale = 1.0)
+void foliumOfDescartes(scope bool delegate(Vec2f) onPointIsContinue, float phi, float step = 0.01, float scale = 1.0)
 {
     assert(onPointIsContinue);
 
     onPointStep(step, 0, Math.PI * 2, (angle) {
         const r = (3 * phi * Math.cos(angle) * Math.sin(angle)) / (
             (Math.cos(angle) ^^ 3) + (Math.sin(angle) ^^ 3)) * scale;
-        return onPointIsContinue(Vec2d.fromPolarRad(angle, r));
+        return onPointIsContinue(Vec2f.fromPolarRad(angle, r));
     });
 }
 
-void tractrix(scope bool delegate(Vec2d) onPointIsContinue, float length, float step = 0.01)
+void tractrix(scope bool delegate(Vec2f) onPointIsContinue, float length, float step = 0.01)
 {
     assert(onPointIsContinue);
 
@@ -120,21 +120,21 @@ void tractrix(scope bool delegate(Vec2d) onPointIsContinue, float length, float 
         const x = sign * length * (
             (log(Math.tan(Math.abs(angle / 2)))) + Math.cos(Math.abs(angle)));
         const y = length * Math.sin(angle);
-        return onPointIsContinue(Vec2d(x, y));
+        return onPointIsContinue(Vec2f(x, y));
     });
 }
 
-void trisectrixMaclaurin(scope bool delegate(Vec2d) onPointIsContinue, float radius = 10, float step = 0.1)
+void trisectrixMaclaurin(scope bool delegate(Vec2f) onPointIsContinue, float radius = 10, float step = 0.1)
 {
     assert(onPointIsContinue);
 
     onPointStep(step, -step, Math.PI * 2, (angle) {
         auto r = (radius / 2) * (4 * Math.cos(angle) - Math.sec(angle));
-        return onPointIsContinue(Vec2d.fromPolarRad(angle, r));
+        return onPointIsContinue(Vec2f.fromPolarRad(angle, r));
     });
 }
 
-void lissajous(scope bool delegate(Vec2d) onPointIsContinue, float amplitudeX = 50, float freqX = 1, float amplitudeY = 50, float freqY = 2, float phaseShift = (
+void lissajous(scope bool delegate(Vec2f) onPointIsContinue, float amplitudeX = 50, float freqX = 1, float amplitudeY = 50, float freqY = 2, float phaseShift = (
         Math.PI / 2), float dots = 2000, float step = 0.01)
 {
     assert(onPointIsContinue);
@@ -145,14 +145,14 @@ void lissajous(scope bool delegate(Vec2d) onPointIsContinue, float amplitudeX = 
         dt += step;
         const x = amplitudeX * Math.sin(freqX * dt + phaseShift);
         const y = amplitudeY * Math.sin(freqY * dt);
-        if (!onPointIsContinue(Vec2d(x, y)))
+        if (!onPointIsContinue(Vec2f(x, y)))
         {
             break;
         }
     }
 }
 
-void rose(scope bool delegate(Vec2d) onPointIsContinue, float roseSize, float n, float d, size_t curlsCount = 1, float step = 0.01)
+void rose(scope bool delegate(Vec2f) onPointIsContinue, float roseSize, float n, float d, size_t curlsCount = 1, float step = 0.01)
 {
     assert(onPointIsContinue);
 
@@ -161,6 +161,6 @@ void rose(scope bool delegate(Vec2d) onPointIsContinue, float roseSize, float n,
     //For an integer k, the number of petals is k if k is odd and 2k if even
     onPointStep(step, 0, Math.PI * curlsCount - step, (angle) {
         auto r = roseSize * Math.sin(petalsFactor * angle);
-        return onPointIsContinue(Vec2d.fromPolarRad(angle, r));
+        return onPointIsContinue(Vec2f.fromPolarRad(angle, r));
     });
 }
